health datatype decimal -> float && fixed directives and ng-controller variables

+++ db/migrate/20140501002649_add_health_to_connections.rb
@@ -1,5 +1,5 @@
 class AddHealthToConnections < ActiveRecord::Migration
   def change
-    add_column :connections, :health, :decimal
+    add_column :connections, :health, :float
   end
 end
